Made running on the UI thread optional
Since anything VS needs to do on the UI thread as a result
of the APIs you invoke can cause it to deadlock waiting for
your test code to finish, it's better to run the tests on a
separate thread, and only opt-in to the UI thread if you really
need it.

This also allows users to switch to using a more sophisticated
way of switching to the UI thread, such as the MS.VS.Threading
library.