Allow customizing all assembly level attribute values from MSBuild
This makes it easier to reuse this codegen by default, and condition
it to environment variables that might be set by a CI system.

Add also a '.' special value for RootSuffix which denotes the normal
hive, which also makes it possible to opt-out of the default behavior
of considering no value or an empty value as a skip for generating the
RootSuffix attribute parameter. This maintains the existing behavior
of always choosing the Exp suffix, while introducing the new value
and allowing more control via MSBuild and automatic codegen.