Remove dependency on VS Shell.11 package, depend on public VSSDK packages
By trimming down the dependencies on the VSSDK shell assemblies, we can
successfully use xunit.vsix with arbitrary test projects that reference
whichever version of the Shell assemblies they want. The dependency was
mainly on the convenience class ServiceProvider, which we now duplicate
(and simplify since we don't use IObjectWithSite anywhere).

Also declare as dependencies of the package the official VSSDK assemblies
for the things we truly need, like EnvDTE and Shell.Interop.