Add support for VS2022 / 64bit
Switch to the injector helper for initial injection.
Remove dependencies on VSSDK interop assemblies and instead just bring in just the absolute minimal set of COM interop interfaces needed to boostrap and run.

Remove the confusing GlobalServices class, which is fully replaceable with the VSSDK's ServiceProvider.GlobalProvider

This also requires dropping support for VS2015 and previous, since there is really no way to have unit tests that can simultaneously target the range of VS versions, VSSDK and installer technologies. This was causing issues for the most part, and by now, I doubt there is much interest in VS2015 and prior versions anyway.

We introduce support for version ranges now, which is way more flexible when authoring tests that can run on a range of installed versions (i.e. 15.*-16.* but not 17.*, would be `[15.0-16.99]` for example). This also allows targetting a very specific VS version (say, `[17.2]`), which can be handy in some cases.

Added unit tests for VsSetup and VsVersions behavior.
The version range matching and evaluation against installed versions is done using NuGet's versioning implementation, so this table exlains the supported syntax and expected behavior: https://docs.microsoft.com/en-us/nuget/concepts/package-versioning#version-ranges