Fix improper defaulting of VsixFact for VS versions
Previously, we'd never pick a class (or assembly) level default VisualStudioVersion because we were defaulting the value in the attribute construction, meaning that property would never be null and therefore the current aggregation behavior we have for other values would never kick in.

This fixes it and makes the current behavior more intuitive: if no specific VS version is specified, the default will instead be the current VS version.