Add assembly metadata-level fallback support
Since we won't always be able to generate the assembly info file, falling back to the SDK built-in codegen for assembly metadata attributes gives us more resiliency in case anything goes wrong in our custom mechanism.

Updated targets to account for the new `*` value that represents "All" VS versions, which was otherwise broken.