Try harder to always have a CurrentVersion
In the scenario where tests are run with dotnet test, none of the VS-specific envvars beyond VisualStudioVersion will be present, so we won't find an exact match, or possibly even a major-only match (if even VisualStudioVersion is missing).

So provide even more defaulting so we can successfully run against something we do find installed.